requirements.txt
Flask==3.0.3
girth==0.8.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
PasteScript==3.6.0
Pillow==10.3.0
scikit_learn==1.4.2
waitress==2.1.2
